style: move style from home to thumbnail container
