+++ src/components/shared/thumbnailContainer/styled.js
